pyproject.toml
[tool.poetry]
name = "ds-project1"
version = "0.1.0"
description = ""
authors = ["ThePsychologist-EV <psychology.data.science@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13"
streamlit = ">=1.53.1,<2.0.0"
pandas-stubs = "~=2.3.3"
matplotlib = "^3.10.8"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
